change all double value to float

+++ Main.c
@@ -5,37 +5,37 @@
 int main()
 {
     int totalDays;
-    double arrivalTime;
-    double departTime;
-    double airFare;
-    double carRent;
-    double miles;
-    double privateCar;
-    double milesDriven;
-    double registrationFee;
-    double confSemFees;
-    double rentalPrice;
-    double rentalValue;
-    double privateValue;
-    double breakfast;
-    double lunch;
-    double dinner;
-    double totalFood;
-    double calcPark;
-    double TotalOverall;
+    float arrivalTime;
+    float departTime;
+    float airFare;
+    float carRent;
+    float miles;
+    float privateCar;
+    float milesDriven;
+    float registrationFee;
+    float confSemFees;
+    float rentalPrice;
+    float rentalValue;
+    float privateValue;
+    float breakfast;
+    float lunch;
+    float dinner;
+    float totalFood;
+    float calcPark;
+    float TotalOverall;
 
     // fees that are partly covered by company
-    double parkingFees;
+    float parkingFees;
     int parkingDays;
-    double taxiFee;
+    float taxiFee;
     int taxiDays;
-    double hotelFee;
+    float hotelFee;
     int hotelDays;  // can be different from totalDays Ex: we can go 4 days and 3 nights
     char carRental[10];
     char taxiUse[10];
     char privateC[10];
-    double minTime = 0.0; //min number is 1
-    double maxTime = 24.0; //max number is 10
+    float minTime = 0.0; //min number is 1
+    float maxTime = 24.0; //max number is 10
     
     printf("Total travel days:");
     scanf("%d", &totalDays);
@@ -49,35 +49,35 @@ int main()
     
     
     printf("\nEnter departure time (0..24 military time):");
-    scanf("%lf", &departTime);
+    scanf("%f", &departTime);
     do{
         if(departTime > maxTime || departTime < minTime ){
             printf("\nTime is out of bound");
             printf("\nEnter departure time (0..24 military time):");
-            scanf("%lf", &departTime);
+            scanf("%f", &departTime);
         }
     }
     while(departTime > maxTime || departTime < minTime);
     
     printf("\nEnter arrival time (0..24 military time):");
-    scanf("%lf", &arrivalTime);
+    scanf("%f", &arrivalTime);
         do{
         if(arrivalTime > maxTime || arrivalTime < minTime){
             printf("\nTime is out of bound");
             printf("\nEnter departure time (0..24 military time):");
-            scanf("%lf", &arrivalTime);
+            scanf("%f", &arrivalTime);
         }
     }
     while(arrivalTime > maxTime || arrivalTime < minTime);
 
     printf("\nAmount of the round trip airfare:$");
-    scanf("%lf", &airFare);
+    scanf("%f", &airFare);
     do{
         if(airFare < minTime)
         {
             printf("\nMoney is never negative");
             printf("\nAmount of the round trip airfare:$");
-            scanf("%lf", &airFare);
+            scanf("%f", &airFare);
         }
     }while(airFare<minTime);
 
@@ -87,12 +87,12 @@ int main()
     if(strcmp(carRental, "y")==0)
     {
         printf("\nHow much was the car you had rented?:$" );
-        scanf("%lf", &rentalPrice);
+        scanf("%f", &rentalPrice);
         do{
             if(rentalPrice<minTime){
                 printf("\nNo such thing as negative money!");
                 printf("\nHow much was the car you had rented?:$" );
-                scanf("%lf", &rentalPrice);
+                scanf("%f", &rentalPrice);
             }
         }while(rentalPrice<minTime);
     }else{
@@ -104,12 +104,12 @@ int main()
     scanf("%s", privateC);
     if(strcmp(privateC, "y")==0){
         printf("\nHow many miles did you drive in your private vehicle? ");        
-        scanf("%lf", &privateCar);
+        scanf("%f", &privateCar);
         do{
             if(privateCar<minTime){
                 printf("\nNo such thing as negative mile!");
                 printf("\nHow many miles did you drive in your private vehicle:" );
-                scanf("%lf", &privateCar);
+                scanf("%f", &privateCar);
             }
         }while(privateCar<minTime);
         privateValue = carMileage(privateCar);
@@ -119,12 +119,12 @@ int main()
     }
     if(rentalPrice || privateCar != 0){
     printf("\nParking fees during trip:$");
-    scanf("%lf", &parkingFees);
+    scanf("%f", &parkingFees);
     do{
         if(parkingFees<minTime){
             printf("\nMoney is never negative");
             printf("\nParking fees during trip:$");
-            scanf("%lf", &parkingFees);
+            scanf("%f", &parkingFees);
         }
         else if(parkingFees > minTime){
             printf("\nAmount of days you used the parking: ");
@@ -143,12 +143,12 @@ int main()
     scanf("%s", taxiUse);
     if(strcmp(taxiUse, "y")==0){
         printf("\nTaxi fees:$");
-        scanf("%lf", &taxiFee);
+        scanf("%f", &taxiFee);
         do{
             if(taxiFee<minTime){// taxiUse -> taxiFee
                 printf("\nFees are not negative");
                 printf("\nTaxi fees:$");
-                scanf("%lf", &taxiFee);
+                scanf("%f", &taxiFee);
             }
             else if(taxiFee > minTime){
                 printf("\nAmount of days used on taxi:");
@@ -168,96 +168,96 @@ int main()
     }
 
     printf("\nConference or seminar registration fees:$");
-    scanf("%lf", &confSemFees);
+    scanf("%f", &confSemFees);
     do{
         if(confSemFees<minTime){
             printf("\nMoney isnt negative!");
             printf("\nConference or seminar registration fees:$");
-            scanf("%lf", &confSemFees);
+            scanf("%f", &confSemFees);
         }
     }while(confSemFees<minTime);
     printf("\nAny hotel expenses:$");
-    scanf("%lf", &hotelFee);
+    scanf("%f", &hotelFee);
     do{
         if(hotelFee<minTime){
             printf("\nMoney isn't negative!");
             printf("\nAny hotel expenses:$");
-            scanf("%lf", &hotelFee);
+            scanf("%f", &hotelFee);
         }
     }while(hotelFee<minTime);
     hotelFee = hotel(hotelFee, totalDays);
     
     printf("\nAmount spent on breakfast:$");
-    scanf("%lf", &breakfast);
+    scanf("%f", &breakfast);
     do{
         if(breakfast<minTime){
             printf("\nAmount spent on breakfast:$");
-            scanf("%lf", &breakfast);
+            scanf("%f", &breakfast);
         }
     }while(breakfast<minTime);
     printf("\nAmount spent on lunch:$");
-    scanf("%lf", &lunch);
+    scanf("%f", &lunch);
     do{
         if(lunch<minTime){
             printf("\nAmount spent on lunch:$");
-            scanf("%lf", &lunch);
+            scanf("%f", &lunch);
         }
     }while(lunch<minTime);
     printf("\nAmount spent on dinner:$");
-    scanf("%lf", &dinner);
+    scanf("%f", &dinner);
     do{
         if(dinner<minTime){
             printf("\nAmount spent on dinner:$");
-            scanf("%lf", &dinner);
+            scanf("%f", &dinner);
         }
     }while(dinner<minTime);
     
     totalFood = breakfast + lunch + dinner;
     calcPark = calcParking(parkingFees, parkingDays);
     TotalOverall = TotalExpense(airFare, rentalPrice, privateValue, calcPark, taxiFee, confSemFees, hotelFee, totalFood);
-    double defaultPark = 6 * totalDays;
-    double defaultTaxi = 10 * totalDays;
-    double defaultHotel = 90 * totalDays;
-    double temp = Totaltemp(airFare, rentalPrice, privateValue, confSemFees);
-    double TotalAllow = temp + defaultPark + defaultHotel + 37;
+    float defaultPark = 6 * totalDays;
+    float defaultTaxi = 10 * totalDays;
+    float defaultHotel = 90 * totalDays;
+    float temp = Totaltemp(airFare, rentalPrice, privateValue, confSemFees);
+    float TotalAllow = temp + defaultPark + defaultHotel + 37;
     printf("\n--------------------------");
     printf("\nTOTAl EXPENSES LIST");
     printf("\n--------------------------");
-    printf("\nTrip:$%lf", airFare);
-    printf("\nRental:$%lf", rentalPrice);
-    printf("\nMileage:$%lf", privateValue );
-    printf("\nParking:$%lf", calcPark);
-    printf("\nTaxi:$%lf", taxiFee);
-    printf("\nConference:$%lf", confSemFees);
-    printf("\nHotels:$%lf", hotelFee);
-    printf("\nBreakfast:$%lf", bf(breakfast, arrivalTime, departTime));
-    printf("\nLunch:$%lf", lch(lunch, arrivalTime, departTime));
-    printf("\nDinner:$%lf",dr(dinner, arrivalTime, departTime));
-    printf("\nTotal Expenses:$%lf", TotalOverall);
+    printf("\nTrip:$%.2f", airFare);
+    printf("\nRental:$%.2f", rentalPrice);
+    printf("\nMileage:$%.2f", privateValue );
+    printf("\nParking:$%.2f", calcPark);
+    printf("\nTaxi:$%.2f", taxiFee);
+    printf("\nConference:$%.2f", confSemFees);
+    printf("\nHotels:$%.2f", hotelFee);
+    printf("\nBreakfast:$%.2f", bf(breakfast, arrivalTime, departTime));
+    printf("\nLunch:$%.2f", lch(lunch, arrivalTime, departTime));
+    printf("\nDinner:$%.2f",dr(dinner, arrivalTime, departTime));
+    printf("\nTotal Expenses:$%.2f", TotalOverall);
     printf("\n--------------------------");
     printf("\nTOTAL ALLOWABLE EXPENSES");
     printf("\n--------------------------");
-    printf("\nTrip:$%lf", airFare);
-    printf("\nRental:$%lf", rentalPrice);
-    printf("\nMileage:$%lf", privateValue );
-    printf("\nParking:$%lf", defaultPark);
-    printf("\nTaxi:$%lf", defaultTaxi);
-    printf("\nConference:$%lf", confSemFees);
-    printf("\nHotels:$%lf", defaultHotel);
+    printf("\nTrip:$%.2f", airFare);
+    printf("\nRental:$%.2f", rentalPrice);
+    printf("\nMileage:$%.2f", privateValue );
+    printf("\nParking:$%.2f", defaultPark);
+    printf("\nTaxi:$%.2f", defaultTaxi);
+    printf("\nConference:$%.2f", confSemFees);
+    printf("\nHotels:$%.2f", defaultHotel);
     printf("\nBreakfast:$9");
     printf("\nLunch:$12");
     printf("\nDinner:$16");
-    printf("\nTotal allowable Expenses:$%lf", TotalAllow);
+    printf("\nTotal allowable Expenses:$%.2f", TotalAllow);
     printf("\n--------------------------");
     printf("\nResult of trip");
     printf("\n--------------------------");
-    double valueCheck = (TotalOverall - TotalAllow);
+    float valueCheck = (TotalOverall - TotalAllow);
     if(valueCheck > 0){
-        printf("\nYou owe the company:$%lf",valueCheck);
+        printf("\nYou owe the company:$%.2f",valueCheck);
     }
     if(valueCheck < 0 ){
-        double conv = abs(valueCheck);
-        printf("\nYou saved the company:$%lf",conv);
+        float conv = abs(valueCheck);
+        printf("\nYou saved the company:$%.2f",conv);
     }
     return 0;
 }
